refactor: Move file to specific chat directory
