+++ lib/pages/offre_emploi/offre_emploi_details_page.dart
@@ -9,7 +9,7 @@ import 'package:pass_emploi_app/models/offre_emploi_details.dart';
 import 'package:pass_emploi_app/network/post_tracking_event_request.dart';
 import 'package:pass_emploi_app/pages/chat/chat_partage_bottom_sheet.dart';
 import 'package:pass_emploi_app/pages/offre_page.dart';
-import 'package:pass_emploi_app/presentation/chat_partage_page_view_model.dart';
+import 'package:pass_emploi_app/presentation/chat/chat_partage_page_view_model.dart';
 import 'package:pass_emploi_app/presentation/offre_emploi/offre_emploi_details_page_view_model.dart';
 import 'package:pass_emploi_app/redux/app_state.dart';
 import 'package:pass_emploi_app/ui/app_colors.dart';
