+++ lib/presentation/chat/chat_partage_page_view_model.dart
