+++ lib/presentation/chat/piece_jointe_view_model.dart
@@ -1,7 +1,7 @@
 import 'package:equatable/equatable.dart';
 import 'package:pass_emploi_app/features/chat/piece_jointe/piece_jointe_actions.dart';
 import 'package:pass_emploi_app/features/chat/piece_jointe/piece_jointe_state.dart';
-import 'package:pass_emploi_app/presentation/chat_item.dart';
+import 'package:pass_emploi_app/presentation/chat/chat_item.dart';
 import 'package:pass_emploi_app/presentation/display_state.dart';
 import 'package:pass_emploi_app/redux/app_state.dart';
 import 'package:redux/redux.dart';
@@ -28,16 +28,11 @@ class PieceJointeViewModel extends Equatable {
 }
 
 DisplayState _displayState(String id, PiecesJointesState piecesJointesState) {
-  final status = piecesJointesState.status[id];
-  switch (status) {
-    case null:
-    case PieceJointeStatus.success:
-      return DisplayState.CONTENT;
-    case PieceJointeStatus.loading:
-      return DisplayState.LOADING;
-    case PieceJointeStatus.failure:
-      return DisplayState.FAILURE;
-    case PieceJointeStatus.unavailable:
-      return DisplayState.EMPTY;
-  }
+  return switch (piecesJointesState.status[id]) {
+    null => DisplayState.CONTENT,
+    PieceJointeStatus.success => DisplayState.CONTENT,
+    PieceJointeStatus.loading => DisplayState.LOADING,
+    PieceJointeStatus.failure => DisplayState.FAILURE,
+    PieceJointeStatus.unavailable => DisplayState.EMPTY,
+  };
 }
